+++ lib/src/navigation_bar_item.dart
@@ -5,11 +5,13 @@ class BottomIndicatorNavigationBarItem {
   final dynamic activeIcon;
   final dynamic label;
   final Color? backgroundColor;
+  final double? iconSize;
 
   BottomIndicatorNavigationBarItem({
     required this.icon,
     this.activeIcon,
     this.label,
     this.backgroundColor = Colors.transparent,
+    this.iconSize,
   });
 }